chore(system-tests-k8s): containerDisk poc (#3481)
**Enhancing Config Disk Logic for Improved Performance**

This PR replaces the existing
[CDI](https://kubevirt.io/user-guide/storage/containerized_data_importer/)
logic, where the configuration disk is pushed to a local S3 bucket and
imported via the CDI controller, with a new approach using
[containerDisk](https://kubevirt.io/user-guide/storage/disks_and_volumes/#containerdisk)
logic. The configuration disk is now wrapped as a container image,
pushed to a local container registry, and pulled by the container
runtime during VM creation.

This change is expected to provide better runtime performance and a more
streamlined setup, especially when running a large number of system
tests in parallel.

A cleaner implementation will follow once statistical data confirms a
significant improvement in performance with these changes. We anticipate
having runtime comparisons with FARM in the coming days to validate the
effectiveness of this approach.