feat(ic-nervous-system-agent): Add `submit_proposal` helper (#3451)
Submitting proposals is a common pattern within the nervous system
integration tests, and we will soon be doing it even more within the SNS
CLI. So it makes sense to have a helper function for it in
`ic-nervous-system-agent`.

There is one nuance to this. `ic_agent::Agent` stores the identity of
the sender within it. But pocketic functions take the sender as an
argument. Previously, in `ic-nervous-system-agent`, we just assumed the
sender would always be the anonymous principal, but obviously this won't
work for submitting proposals. So this PR also introduces a way to
specify the sender for pocketic through the new `PocketIcAgent` type.