fix: EXC-1765: Fix `canister_age` metric (#3443)
The `scheduler_canister_age_rounds` metric tracks the difference between
the current round and the round in which the canister last fully
executed. By default, this last executed round is set to zero for newly
created canisters. Consequently, the `canister_age` metric for these new
canisters abruptly increases to the current round, which can be in the
billions.

This PR addresses this issue by excluding canisters with a zero 'last
executed round' from the `canister_age` metric observations.

Furthermore, this PR increases the upper bound of the metric's range to
5000 rounds.