feat(PocketIC): new library function PocketIc::ingress_status_as (#3492)
This PR splits the function `PocketIc::ingress_status` taking an
optional caller into a pair of functions:
- `PocketIc::ingress_status` taking no caller and returning a more
precise return type;
- `PocketIc::ingress_status_as` taking a non-optional caller and
returning the original return type.

This way, the use cases of just fetching the ingress status and also
checking the caller are properly separated (no need to pass `None` in
the former use case) and the return types are more precise.