chore: Remove stream size tracking (#3382)
We track the size of bytes consumed by guaranteed responses in two
parallel ways:
1. `transient_stream_guaranteed_responses_size_bytes` per canister
2. `guaranteed_responses_size_bytes` in Metadata, a `BTreeMap` keyed
`CanisterId`
Keeping these values in sync is implemented by rather large amount of
code. Specifically `transient_stream_guaranteed_responses_size_bytes` is
loaded in a way different from loading the rest of the fields in
`ReplicatedState`, blocking asynchronous checkpointing.
These fields were introduced when 10GiB streams were planned in the
future. Since then the size limit is lowered to 10MiB x number of
subnets`. Because of this the same mechanism was never introduced for
the Best Effort Messages.
Given the lower memory limit for Streams and complexity of code, it
makes sense to remove the logic all together.