chore: [EXC-1835] Make ic management types private to the execution environment (#3496)
This PR introduces Bazel rules which make the IC management canister
types private to the core crates execution_environment, messaging etc.

All users of the _interface_ are in a temporary whitelist, but are urged
to move to the published version of the types asap. That will be defined
in [cdk-rs](https://github.com/dfinity/cdk-rs) and published to
crates.io.

Context:
https://docs.google.com/document/d/1xvE480Tf61SaJc0AP74_4WDbh01alBU2aQHIRwx3XMw/edit?tab=t.0

---------

Co-authored-by: Michael Weigelt <michael.weigelt@dfinity.com>