test: don't run //rs/tests/cross_chain:ic_xc_cketh_test by default on PRs (#3503)
The P90 duration of the `//rs/tests/cross_chain:ic_xc_cketh_test` [is 7
minutes](https://superset.idx.dfinity.network/explore/?slice_id=48)
which is too long to run on PRs by default. So we tag it as a
`long_test` which only runs this test on pushes to master or when CI
Main is manually triggered on the PR branch or when the PR is labelled
with `CI_ALL_BAZEL_TARGETS`.