feat(fuzzing): Add syscall monitoring to canister sandbox (#3420)
In #2513, we introduced the target
`//rs/execution_environment/fuzz:fuzzer_sandbox` that allows fuzzers to
use cansiter sandbox. In this PR, we improve the library to add syscall
monitoring over the sandbox to mimic a pseudo SELinux setting in a test
environment.

The current approach is simple as in the sandbox panics if it performs a
syscall not present in a static list. But the functionality can be
further improved in future PRs.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>