feat(nervous-system): Add crate candid-utils and add proper Candid service argument validation (#3488)
This PR adds a new crate dedicated to Candid utilities that are likely
to be useful in multiple places in the mono repo.

First, we add `candid_utils::validation::validate_upgrade_args`, which
takes a full Candid service and checks if a given upgrade argument
complies with it, using Candid subtyping. In particular, this function
will soon be required in a tool for submitting upgrade proposals for
SNS-controlled canisters.

| [Next PR](https://github.com/dfinity/ic/pull/3439) >