chore(RUN-1012): infrastructure to increase Wasm64 heap memory size (#3385)
This PR prepares infrastructure to increase Wasm64 heap memory size
allowing Wasm64 canisters to run with different heap memory sizes
alongside the Wasm32 canisters.

Note that this PR is a no-op for now and not changing any functionality,
as the Wasm64 heap memory size is still kept at 4GiB. The intent is that
this only allows that in the future the change can happen seamlessly, by
only changing a constant.