test(Consensus): Add a stub test for vetkd (#3524)
This PR sets up the integration tests for the consensus components of
the vetkeys feature.

For now, it only has a system test that adds a vetkey and checks that
the subnet makes progress.
This is to test, that code changes to be added later are not breaking
the subnet, if a key gets added.

As more features are being implemented, we can expand the test to check
more functionality and add more tests, for now it simply serves a smoke
test.