fix: EXC-1815: Fix sandbox processes metric (#3527)
The monitoring thread may collect information about each sandbox process
for a few seconds. Consequently, metrics scraping might occur during
this collection phase, leading to an underreporting of the number of
sandboxes.

This PR addresses this issue by:

1. Collecting metrics into small vectors of `f64` values. This allows
for more efficient and consistent data gathering.
2. Observing the metrics within a tight loop. This ensures that the
latest and most accurate data is captured.

While this approach may not completely eliminate the possibility of an
incomplete sandbox count, it should significantly reduce the likelihood
of this occurring.