feat: CON-1422 Introduce `VetKdPayload` type (#3458)
This PR introduces a new type `VetKdPayload` which holds completed
agreements on a VetKD request. The subnet may either agree on a
successful response, or agree to reject the request (because it timed
out or requested an invalid key).

In a future PR, this payload (or rather its serialized form) will be
added as a new section of the batch payload.

Additionally, we move functions `iterator_to_bytes` and
`slice_to_messages` from `rs/interfaces` to `rs/types` which makes them
easier to test.