refactor(nns): Delete ManageNeuronResponse from governance.proto. (#3573)
The idea here is very much the same as the [previous PR] in this
thread/series. Therefore, please see the description of the
aforementioned PR for an overview, and details.

[previous PR]:https://github.com/dfinity/ic/pull/3546

One difference here is that we only deleted the response type, not the
request type. In this case, deleting the request type is not possible,
because it is used (indirectly) by `Governance` (via `Proposal`). This
is an unusual situation; most request types do not have this property.