chore: remove `ic_api_version` field in IC status response (#3569)
This PR removes the `ic_api_version` field in `HttpStatusResponse` for
the http status endpoint of the IC. The field is never updated, nor
populated with the right value, meaning it adds no value. NET-776 aimed
to use env variables for the value, but was completed as won't do.

The `ic_api_version` field is not used in the agents.