test: Add best effort messages to random traffic canister (#3108)
Adds support for best-effort calls on the random traffic canister. All
arguments used to run the canister are collected in the `Config` now
(except the random seed) since otherwise the number of update calls and
individual parameters would get quite large.
Adds a new test where a local canister makes only best-effort calls, to
itself and to a remote subnet. Checks that the local canister can be
stopped even if the remote subnet stalls.

To change any parameter you can query the current config, then change it
accordingly and send it back to the canister. The seed is excluded since
reseeding the rng every time would reset the random sequence.

Additionally the records now contain a timeout if the call is
best-effort and the time from making the call to receiving a reply is
measured (for all calls).

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>