docs(governance): Added 2025-01-20 CHANGELOG.md entries. (#3535)
Used the new add-release-to-changelog script.

Needed to hand craft the "Features & Fixes" section in some proposals,
because we forgot to update unreleased_changelog.md, most likely due to
this being a new process. Other than that, I think the script worked.

Needed to move registry changelog files to the the canister subdir.

# Canisters & Proposals

* NNS
    * registry - https://dashboard.internetcomputer.org/proposal/134904

* SNS
    * root - https://dashboard.internetcomputer.org/proposal/134905
* governance - https://dashboard.internetcomputer.org/proposal/134906
    * swap - https://dashboard.internetcomputer.org/proposal/134907