fix: use api boundary nodes as socks proxies (#2712)
https://github.com/dfinity/ic/pull/2191 creates socks proxies on API
Boundary Nodes.

This PR fetches all of the API Boundary Nodes IPs through registry for
each https outcalls request batch and sends them to the adapter. The
adapter uses one of these IPs as SOCKS Proxy (if needed), at random,
with a maximum of 3 retries.

This is a **dark launch**; still using the old boundary nodes as socks
proxies, and comparing the results with results coming through the new
api boundary nodes.

As bandwidth could go up drastically, we only perform dark launch
comparisons on a small percentage of requests. That percentage is
currently set to 0.