chore: add V3 to ICRC Ledger canister revisions and update mainnet to V4 (#3570)
V3 was added and mainnet changed to V4. The tests that require more than
one migration steps cannot be performed with mainnet anymore (no
migration happens) and were changed to V3. Upgrading from u64 to u256 is
not possible anymore, however it is not caught by the upgrade. In order
to enable this test we need to add additional check to mainnet
canisters.