feat(rosetta): Local cluster setup for Rosetta (#3485)
This PR introduces a script that manages a local kubernetes set up that
runs Rosetta (icp and icrc) with prod images as well as locally built
images.
The cluster also provides a Grafana instance connected to the rosetta
nodes that allows us to easily track system utilization (CPU, Memory,
IO) as well as other custom metrics.

The intent is to have a quick way to investigate issues and visualize
the impact of local changes against the prod versions.

See the README.md file for more details.