feat: [MR-354] Load & validate protos asynchronously (#2594)
As a part of the asynchronous checkpointing, we move reading protos to
the Tip thread.
Currently we have the ReplicatedState from execution, serialize it into
a checkpoint, load back the checkpoint in order to ensure it's loadable
and equivalent to the original one, use the loaded checkpoint to switch
the file descriptors in the original state to point to the new files.
The protos are only used for validation and at replica init.
With this change we do loading and validation away from the critical
path.

---------

Co-authored-by: Stefan Schneider <31004026+schneiderstefan@users.noreply.github.com>