feat(cycles-minting): Cycles Minting canister refunds automatically. (#3484)
This happens when the `memo` (or `icrc1_memo`) in an ICP transfer is not
one of the special values that CMC recognizes as signaling the purpose
of the transfer.

This behavior gets triggered when one of the `notify_*` methods gets
called.

The new behavior is not actually active yet. To turn it on, we simply
need to set `IS_AUTOMATIC_REFUND_ENABLED` to true. I have run the tests
with this set to true locally, and they pass.

# Motivation

Avoids ICP getting stuck, which can result from people making mistakes
when sending ICP to the CMC.

# Implementation Overview

Added `issue_automatic_refund_if_memo_not_offerred`. As indicated by the
name, this is the heart of the implementation of this new functionality.
This is called from `fetch_transaction`. The call is guarded behind the
aforementioned `IS_AUTOMATIC_REFUND_ENABLED` flag.

Added `MEANINGFUL_MEMOS` constant. This is used to detect when automatic
refund should be performed.

# Minor Changes

Added `AutomaticallyRefunded` to `NotificationStatus`. This ensures that
we do NOT duplicate refunds.

Changed one parameter of `fetch_transaction` from `AccountIdentifier` to
`Subaccount`, because you cannot go "backwards" from `AccountIdentifier`
to `Subaccount`.

Factored out the `memo` + `icrc1_memo` unifying code which used to live
in `transaction_has_expected_memo`. Now, that code lives in a new
`get_u64_memo` function.

# Future Work

Added a couple of other helpers. One is
`set_block_status_to_processing`. This could be used to reduce code
duplication in `notify_*` methods, but such refactoring is left to
future PRs.

# References

We promised to do [this in this forum post][promise].

[promise]:
https://forum.dfinity.org/t/extend-cycles-minting-canister-functionality/37749/2?u=daniel-wong

The other feature that we promised in that thread (use icrc1_memo) was
implemented in [an earlier PR][icrc1_memo].

[icrc1_memo]: https://github.com/dfinity/ic/pull/3336