fix(nns): Fix neurons_fund_total_active_neurons metric (#3610)
# Why

The `neurons_fund_total_active_neurons` calculation was wrong when
`allow_active_neurons_in_stable_memory = true` and
`migrate_active_neurons_to_stable_memory = false` (the first phase of
the migration). This was not detected in tests because the 2 flags
always have the same value (both true or both false).

# What

- Fix the metric calculation
- Add an inactive neuron into test, and test the metrics calculation
under different combinations of the flag
- Documented that the previous calculation of `dissolved_neurons_count`
and `neurons_with_less_than_6_months_dissolve_delay_count` are wrong
when inactive neurons are in stable memory, and letting the tests
reflect that