fix(IDX): don't use interpolation in x86-darwin workflow env
This ensures that the environment variables as set with `env:` do not
use e.g. `${ CI_RUN_ID }`. The values are set as-is, without actual
(bash-style) interpolation.

This caused the `--remap-path-prefix` not to kick in (potentially
ruining cache hits) and caused some build failures (where one cargo
crate did re-evaluate the env var string which _did_ interpolate the
variable in the build, pointing the "build dir" to an incorrect location
(`"/var/tmp/12345/"` instead of `"/var/tmp/${CI_RUN_DIR}"`,
`${CI_RUN_DIR}` verbatime) and failing the build.