feat: ICP-ledger: FI-1440: Implement V4 for ICP ledger - migrate balances to stable structures (#3314)
This version of the ledger migrates the balances stored by the ledger
from heap to stable memory stable structures. This allows the ledger to
store more data (heap memory is limited to 4GB) and reduces the number
of instructions used during the upgrades - there is no need to serialize
and deserialize the balances stored in heap memory. Since we are no
longer limited by the heap size, the logic for trimming balances and
related init/upgrade arguments (`maximum_number_of_accounts`,
`accounts_overflow_trim_quantity`) were removed. The migration to stable
structures is performed during the upgrade and, if 250B instruction
limit is reached, using timer invocations after the upgrade. If timer
invocations are used by the migration, the ledger is unavailable for
transfers and approvals until migration is finished. The migration
status can be checked by calling the `is_ledger_ready` endpoint. After
migration is finished, it is no longer possible to downgrade the ledger
to the previous version .

---------

Co-authored-by: Mathias Björkqvist <mathias.bjorkqvist@dfinity.org>
Co-authored-by: mraszyk <31483726+mraszyk@users.noreply.github.com>