chore(EXC-1840): Refactor bool  to use WasmExecutionMode enum. (#3586)
In the `ExecutionState` we were using a `bool` to represent whether a
certain canister was executing in Wasm64 mode or not. In other places,
such as the `CyclesAccountManager` we were using an enum that was cast
from the bool. This PR refactors the usage of the `bool` parameter to
using the `WasmExecutionMode` enum in all places. The only exception is
the protobufs that are used for storing state on disk. For simplicity
and compatibility we keep the bool representation there and we convert
to the enum on serialization/deserialization.