chore: Reduce flakiness of state sync system tests
This commit fixes a condition that can lead to flakiness in the state
sync system tests.

The test relies on a metric of the number of successful state syncs.
However, this metric is only incremented when the `IncompletState` is
dropped, which is part of the wind down of a state sync. However, it is
possible for the state manager to rely on the downloaded state before
that. This race condition confused the test.

Now, the test waits for the metric to increase instead of asserting on
the first try.