chore: Add dfx-core dependency and use it in the SNS CLI to allow it to use the user's DFX identities (#2927)
This is needed for the DFX SNS extension to better integrate with DFX.
The dependency is maintained by DFINITY, it is just not part of the
monrepo (and not currently depended on by the monorepo, prior to this
PR).

dfx-core uses a newer version of ic-agent, which means we need to update
ic-agent in the monorepo, and that means we need to update
ic-http-gateway (which I do here).

SNS CLI commands after this PR have a slightly different command after
this. You should pass `--network` to specify the network and
`--identity` to specify the identity if you don't want to use the
default one in DFX. The default network is the local one if none is
specified, for compatibility with DFX.