chore(fuzzing): extend syscall whitelist in sandbox fuzzers (#3659)
In the sandbox monitor function, the parent thread calls
`std::thread::sleep(std::time::Duration::from_secs(1))` to allow the
sandbox to spawn all its threads before tracing begins.

However, this approach is unreliable—the sleep duration is arbitrary,
and attachment can occur at any point in execution. This led to
`getrandom` failures in CI.

This PR removes the sleep timer and instead continuously fetches `tid`s
as threads spawn, attaching to them dynamically. The trade-off is that
it requires knowing the number of sandbox threads in advance.