chore: Reduce flakiness of state sync system tests (#3672)
This commit fixes a condition that can lead to flakiness in the state
sync system tests.

The test relies on a metric of the number of successful state syncs.
However, there appears to be a race condition of when the changed metric
can be observed. At the same time, in this test it only takes a very
short amount of time between the end of state sync and the subnet being
no longer stalled. This race condition confused the test.

Now, the test waits for the metric to increase instead of asserting on
the first try.