test(sns-cli): Port UpgradeSnsControlledCanister with Large Wasm integration test to use SNS CLI with PocketIc (#3696)
This PR leverages the SNS CLI functionality to prepare and submit
UpgradeSnsControlledCanister proposals for large Wasms as part of the
existing integration test scenario. The goal is to have the CLI
functionality tested in CI. This entails the need to generalize the CLI
command upgrade-sns-controlled-canister to work with a CallCanister
instance which can be instantiated not only with ic-agent, but also with
PocketIc instances.