sns-testing: use NNS Governance test version
Problem: We want to be able to run pocket-ic in live mode. However, this
is not viable with non-test version of NNS Governance canister because
it postpones the SNS swap start.

Solution: Use test version of NNS Governance canister. Ensure that swap
start_time is 'None' to start the swap immediately after SNS creation
proposal adoption.