refactor(registry): Migrate Registry to use MemoryManager instead of raw stable memory (#3700)
## What
Use MemoryManager instead of dfn_core's memory tools for saving and
loading the protobuf on upgrades.

## Why
To allow moving registry records into stable memory, so that registry
can scale, we first want to allow for multiple memory regions to be
available so that different classes of objects can be stored.

Note, this was tested using the golden_state test and adding an
additional check that registry state is still intact.