refactor(nns-tools): Turn release-runscript to clap-powered CLI (#3712)
This will make it simpler to add automation for various steps in the
future. It will also enable the person doing the release to start at any
step (instead of needing to start at the beginning).

As a reminder, the runscript can be run with:

```
bazel run //rs/nervous_system/tools/release-runscript --config=lint
```

In this chain of PRs, I automate the first few steps of the release
process in this script.

[Next PR →](https://github.com/dfinity/ic/pull/3713)