feat(nns): Bump neurons limit (#3739)
The number of neurons keeps growing. Some analysis showed that we would
be able to increase the limit to 500K (note that the maximum number
changes over time because it takes into account the number of neurons
that are already created). However, given that we are actively working
on moving all neurons to stable storage, we would like to increase the
limit in a more conservative way. The 400K limit should give us ~10
weeks.