nervous_system::intergration_tests:: use ic_nervous_agent in helpers
We want to be able to reuse some of these helpers outside of PocketIC
environment using `ic-agent` as a backend. So we're porting some of the
helpers for NNS/SNS to use 'ic_nervous_agent' instead of 'pocket_ic'.
Note that at the moment these helpers still provide 'pocket_ic'-based
interface and they will be abstracted out later.